% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdlinear.R
\name{hdlinear}
\alias{hdlinear}
\title{Summary of lm and Anova}
\usage{
hdlinear(y, x, data)
}
\arguments{
\item{y}{integer: the colnum number of a dependent variable}

\item{x}{vector for integer: the colnum numbers of independent variables}

\item{data}{an optional data frame containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called}
}
\value{
This function returns a quantile of residuals, coefficients, residaul standard error, R.squared, F.statistic, and analysis of variance table.

t-test and f-test can be done with the results of a function.
}
\description{
hdlinear can be used to carry out the summary of fitted linear models and print anova tables with a large amount of data in Rhadoop
}
\examples{
library(rhdfs)
hdfs.init()
library(rmr2)
rmr.options(backend = "hadoop")
hdlinear(y=5, x=c(7,11), bigdata)
}
